moved test to own package
